+++ src/harlequin/__init__.py
@@ -1,6 +1,7 @@
 from harlequin.adapter import HarlequinAdapter, HarlequinConnection, HarlequinCursor
 from harlequin.app import Harlequin
 from harlequin.autocomplete import HarlequinCompletion
+from harlequin.keymap import HarlequinKeyBinding, HarlequinKeyMap
 from harlequin.options import HarlequinAdapterOption, HarlequinCopyFormat
 from harlequin.transaction_mode import HarlequinTransactionMode
 
@@ -13,4 +14,6 @@ __all__ = [
     "HarlequinCopyFormat",
     "HarlequinCursor",
     "HarlequinTransactionMode",
+    "HarlequinKeyMap",
+    "HarlequinKeyBinding",
 ]