+++ src/harlequin/components/__init__.py
@@ -1,5 +1,5 @@
 from harlequin.components.code_editor import CodeEditor, EditorCollection
-from harlequin.components.data_catalog import DataCatalog
+from harlequin.components.data_catalog import DataCatalog, HarlequinTree
 from harlequin.components.error_modal import ErrorModal
 from harlequin.components.export_screen import ExportScreen, export_callback
 from harlequin.components.help_screen import HelpScreen
@@ -17,5 +17,6 @@ __all__ = [
     "ResultsViewer",
     "RunQueryBar",
     "DataCatalog",
+    "HarlequinTree",
     "HistoryScreen",
 ]